renamed distributions module to soft_labelling
