+++ dlordinal/soft_labelling/general_triangular_distribution.py
