+++ dlordinal/soft_labelling/tests/test_triangular_distribution.py
