+++ dlordinal/soft_labelling/triangular_distribution.py
